LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-no-content-value-expected.txt


FAIL Meta color-scheme without content attribute has no effect. assert_equals: expected "normal" but got "auto"
